Prevent building in bind-mounted workspace folder
Running jack in, building and reloading was super slow as the default code thrown
into the workspace folder from vscode is bind mounted. This runs the
workshop from the /root/cljs-workshop folder instead. Code changes
are not visible outside of the dev container, but at least its fast.

overrideCommand true makes the devcontainer ignore the CMD of the docker
image, and just load it as a base